amd/common: add nir->llvm translation. (v1.1)
This adds the basic files for the NIR->LLVM translation layer,
along with some hopefully generic code to load the binary
result, and other helpers required.

The hope is in the future we could share this with an
GL_ARB_spirv implementation or a push to replace TGSI
with NIR in radeonsi.

v1.1: cleanup whitespace, move Makefiles to separate patch
for now, add missing copyrights, use sid.h, drop
ac_radeon_winsys.h

Authors: Bas Nieuwenhuizen and Dave Airlie
Signed-off-by: Dave Airlie <airlied@redhat.com>