radv: add scratch support for spilling.
Currently LLVM 5.0 has support for spilling to a place
pointed to by the user sgprs instead of using relocations.

This is enabled by using the amdgcn-mesa-mesa3d triple.

For compute gfx shaders we spill to a buffer pointed to
by 64-bit address stored in sgprs 0/1.
For other gfx shaders we spill to a buffer pointed to by
the first two dwords of the buffer pointed to in sgprs 0/1.

This patch enables radv to use the llvm support when present.

This fixes Sascha Willems computeshader demo first screen,
and a bunch of CTS tests now pass.

This patch is likely to be in LLVM 4.0 release as well
(fingers crossed) in which case we need to adjust the detection
logic.

SIgned-off-by: Dave Airlie <airlied@redhat.com>