[rfc] radv: add initial prime support.
This is kind of a gross hacks, but vulkan doesn't specify anything
but it would be nice to let people with prime systems at least
see some stuff rendering for now.

This creates a linear shadow image in GART that gets blitted to at the
image transition.

Now ideally:
this would use SDMA - but we want to use SDMA for transfer queues
maybe we don't expose a transfer queue on prime cards who knows.

we wouldn't have to add two pointers to every image, but my other
attempts at this were ugly.

Is the image transition the proper place to hack this in? not
really sure anywhere else is appropriate.

It also relies on DRI_PRIME=1 being set, I should be able
to work this out somehow automatically I think, probably getting
a DRI3 fd from the X server and doing drmGetDevice on it, and
comparing where we end up.

Signed-off-by: Dave Airlie <airlied@redhat.com>