spirv: add plumbing for SPV_AMD_shader_ballot and Groups
glslang enables the Group capability in SPIR-V when compiling shaders
with GL_AMD_shader_ballot, and uses OpGroupIAdd etc. for implementing
some functions in GL_AMD_shader_ballot, so it seems that the Group
capability is implicitly added by the Vulkan extension. I've added
support for both sets of instructions at the same time since they're
fairly closely related, and handled with the same macro. For the same
reason, I've put the implementation of the OpGroup* instructions in
vtn_amd.c, even though they're not part of an AMD extension.