nir: rename and generalize nir_lower_read_invocation_to_scalar
We'll want to scalarize other intrinsics in a similar manner for
AMD_shader_ballot, and possibly other extensions in the future. This
patch reworks the pass to use the intrinsic's info to detect whether we
need to copy the source or scalarize it, similarly to how ALU operations
are handled, and makes sure we copy indices if there are any. We don't
yet handle any other intrinsics, though, so this should shouldn't have
any functional effect yet.