radv: optimise draw path cpu usage.
This is probably bordering on OCD, and I've used perf
and objdump more than I should.

This splits the draw path into basic and extended paths,
tries to use register to pass parameters instead
of having to shove them all on the stack in the struct,
it keeps using a struct for indirect draws.

fixup