nak/assign_regs: attempt to track vectors across blocks better
This is probably a bit too naive, but needs some guidance.

The idea is to gather all uses of SSA values in vectors across
all blocks, then when checking if an SSA value should be vector
to look it up in the blocks first, I'm not sure ordering here
really matters, as if it's used in a vector at all we might
want to maintain that.

One big change is to ignore phi webs for vectors.

for my one shader.

-Instruction count: 1011
-Static cycle count: 2944
+Instruction count: 923
+Static cycle count: 2624